# Conflicts:
#	stp-ms-lessons/src/main/java/com/sytoss/lessons/services/ExamService.java
#	stp-ms-lessons/src/test/java/com/sytoss/lessons/services/ExamServiceTest.java

+++ stp-domain/src/main/java/com/sytoss/domain/bom/personalexam/PersonalExam.java
@@ -67,7 +67,7 @@ public class PersonalExam {
     private PersonalExamStatus status = PersonalExamStatus.NOT_STARTED;
 
     @JsonView({PersonalExam.Public.class, PersonalExam.TeacherOnly.class})
-    private float summaryGrade;
+    private double summaryGrade;
 
     @JsonView({PersonalExam.Public.class, PersonalExam.TeacherOnly.class})
     private double maxGrade;
